+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to Dimensions will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.17] - 2026-01-03
+
+### Added
+- Jumping to a tab based on the dimension you are hovering on in TUI via ":#"
+- When a session is running, gives small preview of tab
+
 ## [0.2.16] - 2026-01-03
 
 ### Changed
@@ -197,6 +203,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `x` - Remove tab
 - `q` - Quit and detach from tmux
 
+[0.2.17]: https://github.com/KarlVM12/Dimensions/compare/v0.2.16...v0.2.17
 [0.2.16]: https://github.com/KarlVM12/Dimensions/compare/v0.2.15...v0.2.16
 [0.2.15]: https://github.com/KarlVM12/Dimensions/compare/v0.2.14...v0.2.15
 [0.2.14]: https://github.com/KarlVM12/Dimensions/compare/v0.2.13...v0.2.14